+++ pulsequantum/pulsebuilding.py
@@ -1,4 +1,8 @@
 import broadbean as bb
+from os import listdir
+from os.path import isfile, join
+from pathlib import Path
+import pathlib
 from PyQt5.QtCore import QCoreApplication,Qt
 from PyQt5.QtWidgets import QFileDialog
 from PyQt5.QtWidgets import QApplication, QWidget, QFrame,QMainWindow, QPushButton, QAction, QMessageBox, QLineEdit, QLabel, QSizePolicy
@@ -17,9 +21,11 @@ divch=[divch1,divch2,divch3,divch4];
 corrDflag=0; #Global flag: Is correction D pulse already defined in the pulse table?
 
 class Gelem():
-    def __init__(self, AWG=None, gelem=None):
+    def __init__(self, AWG=None, gelem=None, libpath = 'pulselib/'):
         self.gelem = bb.Element()
         self.awgclock=1.2e9
+        self.libpath = join(pathlib.Path(__file__).parents[0], libpath)
+        self.seq_files = [f for f in listdir(self.libpath) if isfile(join(self.libpath, f))]
 
 
     def generateElement(self,table):
@@ -104,11 +110,110 @@ class Gelem():
    
     #def loadElement(self,path):
      #   seq = bb.Sequence.init_from_json(path)
-        
-        
-    def saveElement(self):
-        return None
+    
+    
 
+    def write_element(self,element, path:str,SR:float = 1e9,SeqAmp:float = 10e-3,SeqOffset:float = 0):# -> None
+        if element.SR == None:
+            element.setSR(SR)
+        seqtmp = bb.Sequence()
+        for ch in element.channels:
+            seqtmp.addElement(ch, element)
+            seqtmp.setChannelAmplitude(ch, SeqAmp)
+            seqtmp.setChannelOffset(ch, 0)
+        seqtmp.setSR(element.SR)
+        seqtmp.write_to_json(path)
+
+    def saveElement(self,path):
+        #self.gelem.write_to_json(path)
+        self.write_element(self.gelem,path)
+        self.seq_files = [f for f in listdir(self.libpath) if isfile(join(self.libpath, f))]  
+
+
+        # From Sequence
+    def from_sequence(self, table, seq):
+         
+        seq_description = seq.description['1']['channels']
+        seg_name = []
+        seg_durations = []
+        seg_ramp = []
+        values = []
+        marker1 = []
+        marker2 = []
+        for chan in seq_description.keys():
+            ch_values = []
+            channels_marker1 = []
+            channels_marker2 = []
+            marker1_rel = seq_description[chan]['marker1_rel']
+            marker2_rel = seq_description[chan]['marker2_rel']
+            seg_mar_list = list(seq_description[chan].keys())
+            seg_list = [s for s in seg_mar_list if 'segment' in s]
+            for i, seg in enumerate(seg_list):
+                seg_digt = seq_description[chan][seg]
+                tmp_name = seg_digt['name']
+                tmp_durations = seg_digt["durations"]
+                if tmp_name not in seg_name:
+                    seg_name.append(tmp_name)
+                    seg_durations.append(tmp_durations)
+                    if seg_digt['arguments']['start'] != seg_digt['arguments']['stop']:
+                        seg_ramp.append(1)
+                    else:
+                        seg_ramp.append(0)
+                ch_values.append(seg_digt['arguments']['stop'])
+                if marker1_rel[i] == (0,0):
+                    channels_marker1.append(0)
+                else:
+                    channels_marker1.append(1)
+                    
+                if marker2_rel[i] == (0,0):
+                    channels_marker2.append(0)
+                else:
+                    channels_marker2.append(1)             
+            values.append(ch_values)
+            marker1.append(channels_marker1)
+            marker2.append(channels_marker2)
+         
+        self.nchans = len(values)
+        nsegs = len(values[0])
+
+
+        table.setColumnCount((self.nchans*3)+2)
+        table.setRowCount(nsegs)
+        
+        #Set horizontal headers
+        h=self.nchans+1;
+        table.setHorizontalHeaderItem(0, QTableWidgetItem("Time (us)"));
+        table.setHorizontalHeaderItem(1, QTableWidgetItem("Ramp? 1=Yes"));
+        for i in range(self.nchans):
+            table.setHorizontalHeaderItem(i+2, QTableWidgetItem("CH%d"%(i+1)));
+            table.setHorizontalHeaderItem(h+1, QTableWidgetItem("CH%dM1"%(i+1)));
+            table.setHorizontalHeaderItem(h+2, QTableWidgetItem("CH%dM2"%(i+1)));
+            h=h+2;
+        
+        #Set vertical headers
+        #nlist= seg_name
+        for i, name in enumerate(seg_name):
+            table.setVerticalHeaderItem(i, QTableWidgetItem(name));
+            
+        
+        for seg in range(nsegs):
+            duration = str(seg_durations[seg]/1e-6)
+            table.setItem(seg,0, QTableWidgetItem(duration))
+            ramp_yes = str(seg_ramp[seg])
+            table.setItem(seg,0, QTableWidgetItem(duration))
+            table.setItem(seg,1, QTableWidgetItem(ramp_yes))
+            for ch in range(self.nchans):
+               val = str(values[ch][seg]/(divch[ch]*1e-3))
+               mark1 = str(marker1[ch][seg])
+               mark2 = str(marker2[ch][seg])
+               table.setItem(seg,ch+2, QTableWidgetItem(val))
+               table.setItem(seg,ch*2+4, QTableWidgetItem(mark1))
+               table.setItem(seg,ch*2+5, QTableWidgetItem(mark2))
+
+    def loadElement(self,table, path):
+        seq = bb.Sequence.init_from_json(path)
+        self.from_sequence(table, seq=seq)
+    
         
     def plotElement(self):
         plotter(self.gelem)
